requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
SQLAlchemy==1.4.54
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.9